# Backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Test API with swagger",
    "version": "1.0.0",
    "description": ""
  },
  "basePath": "/api",
...
  "schemes": [
    "http"
  ],
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/quiz/create": {
      "post": {
        "tags": [
          "Quiz"
        ],
        "description": "status는 pending/open/done으로 나뉨",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "type": "object",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "중간점검 OX"
                },
                "subjectId": {
                  "type": "number",
                  "example": 0
                },
                "answerSheets": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "question": {
                        "type": "string",
                        "example": "배고파?"
                      },
                      "answer": {
                        "type": "string",
                        "example": "O"
                      }
                    },
                    "required": [
                      "question",
                      "answer"
                    ]
                  }
                },
                "type": {
                  "type": "string",
                  "example": "ox"
                }
              },
              "required": [
                "name",
                "subjectId",
                "answerSheets",
                "type"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "퀴즈가 정상적으로 생성된 경우",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "quiz": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "number",
                      "example": 0
                    },
                    "name": {
                      "type": "string",
                      "example": "중간점검 OX"
                    },
                    "subjectId": {
                      "type": "number",
                      "example": 0
                    },
                    "date": {
                      "type": "string",
                      "example": ""
                    },
                    "deadLine": {
                      "type": "string",
                      "example": ""
                    },
                    "answerSheets": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "question": {
                            "type": "string",
                            "example": "배고파?"
                          },
                          "answer": {
                            "type": "string",
                            "example": "O"
                          },
                          "points": {
                            "type": "number",
                            "example": 5
                          }
                        },
                        "required": [
                          "question",
                          "answer",
                          "points"
                        ]
                      }
                    },
                    "status": {
                      "type": "string",
                      "example": "pending"
                    },
                    "type": {
                      "type": "string",
                      "example": "ox"
                    },
                    "responses": {
                      "type": "array",
                      "example": [],
                      "items": {}
                    }
                  },
                  "required": [
                    "answerSheets",
                    "status",
                    "type"
                  ]
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "user가 로그인이 되지 않은 경우",
            "schema": {
              "$ref": "#/definitions/authFailed"
            }
          },
          "403": {
            "description": "type이 professor가 아닌 경우",
            "schema": {
              "$ref": "#/definitions/proAuthFailed"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/quiz/get/subject/{id}": {
      "get": {
        "tags": [
          "Quiz"
        ],
        "description": "status는 pending/open/done으로 나뉨",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "해당 과목에 속하는 quiz 전부 정상적으로 받아온 경우",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "quizzes": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "number",
                        "example": 0
                      },
                      "name": {
                        "type": "string",
                        "example": "중간점검 OX"
                      },
                      "subjectId": {
                        "type": "number",
                        "example": 0
                      },
                      "date": {
                        "type": "string",
                        "example": ""
                      },
                      "deadLine": {
                        "type": "string",
                        "example": ""
                      },
                      "answerSheets": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "question": {
                              "type": "string",
                              "example": "배고파?"
                            },
                            "answer": {
                              "type": "string",
                              "example": "O"
                            },
                            "points": {
                              "type": "number",
                              "example": 5
                            }
                          },
                          "required": [
                            "question",
                            "answer",
                            "points"
                          ]
                        }
                      },
                      "status": {
                        "type": "string",
                        "example": "pending"
                      },
                      "type": {
                        "type": "string",
                        "example": "ox"
                      },
                      "responses": {
                        "type": "array",
                        "example": [],
                        "items": {}
                      }
                    },
                    "required": [
                      "answerSheets",
                      "status",
                      "type"
                    ]
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "user가 로그인이 되지 않은 경우",
            "schema": {
              "$ref": "#/definitions/authFailed"
            }
          },
          "403": {
            "description": "type이 professor가 아닌 경우",
            "schema": {
              "$ref": "#/definitions/proAuthFailed"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/quiz/get/student/subject/{id}": {
      "get": {
        "tags": [
          "Quiz"
        ],
        "description": "status는 pending/open/done으로 나뉨",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "정상적으로 학생이 열람가능한 퀴즈들에서 본인과 관련된 정보만을 받아왔을 경우",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "quizzes": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "quizId": {
                        "type": "number",
                        "example": 0
                      },
                      "name": {
                        "type": "string",
                        "example": "중간점검 OX"
                      },
                      "subjectId": {
                        "type": "number",
                        "example": 0
                      },
                      "questions": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "question": {
                              "type": "string",
                              "example": "배고파?"
                            },
                            "points": {
                              "type": "number",
                              "example": 5
                            }
                          },
                          "required": [
                            "question",
                            "points"
                          ]
                        }
                      },
                      "date": {
                        "type": "string",
                        "example": "2021-05-05T15:38:19.424Z"
                      },
                      "deadLine": {
                        "type": "string",
                        "example": "2021-05-05T16:00:00.000Z"
                      },
                      "status": {
                        "type": "string",
                        "example": "open"
                      },
                      "type": {
                        "type": "string",
                        "example": "ox"
                      },
                      "responses": {
                        "type": "array",
                        "example": [],
                        "items": {}
                      }
                    },
                    "required": [
                      "questions",
                      "status",
                      "type"
                    ]
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "user가 로그인이 되지 않은 경우",
            "schema": {
              "$ref": "#/definitions/authFailed"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/quiz/get/student/{id}": {
      "get": {
        "tags": [
          "Quiz"
        ],
        "description": "status는 pending/open/done으로 나뉨",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "해당 퀴즈에서 정상적으로 본인과 관련된 정보만을 받아왔을 경우",
            "schema": {
              "type": "object",
              "properties": {
                "quizId": {
                  "type": "number",
                  "example": 0
                },
                "name": {
                  "type": "string",
                  "example": "중간점검 OX"
                },
                "subjectId": {
                  "type": "number",
                  "example": 0
                },
                "questions": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "question": {
                        "type": "string",
                        "example": "배고파?"
                      },
                      "points": {
                        "type": "number",
                        "example": 5
                      }
                    },
                    "required": [
                      "question",
                      "points"
                    ]
                  }
                },
                "date": {
                  "type": "string",
                  "example": "2021-05-05T15:38:19.424Z"
                },
                "deadLine": {
                  "type": "string",
                  "example": "2021-05-05T16:00:00.000Z"
                },
                "status": {
                  "type": "string",
                  "example": "open"
                },
                "type": {
                  "type": "string",
                  "example": "ox"
                },
                "responses": {
                  "type": "array",
                  "example": [],
                  "items": {}
                }
              },
              "required": [
                "questions",
                "status",
                "type"
              ],
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "user가 로그인이 되지 않은 경우",
            "schema": {
              "$ref": "#/definitions/authFailed"
            }
          },
          "409": {
            "description": "해당 quiz의 status가 학생이 열람하기에 허용되지 않은 경우",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "isPending": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "authFailed": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": false
        },
        "isLogined": {
          "type": "boolean",
          "example": false
        }
      }
    },
    "proAuthFailed": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": false
        },
        "isProfessor": {
          "type": "boolean",
          "example": false
        }
      }
    }
  }
}